Refactor drag and gravity calculations with velocity threshhold

+++ src/Movables/Circle.java
@@ -10,7 +10,7 @@ public class Circle extends Movable {
 
     public double radius;
 
-    public Circle(double radius, Pair<Integer, Integer> pos, double mass, double terminalVelocity, double elasticity, Color color) {
+    public Circle(double radius, Pair<Double, Double> pos, double mass, double terminalVelocity, double elasticity, Color color) {
         super(pos, mass, terminalVelocity, elasticity, color);
         this.radius = radius;
     }
